server : reorganize structs and enums + naming fixes
